requirements.txt
httpx==0.26.0
httpx-cache==0.13.0
jsonpath-ng==1.6.1
loguru==0.5.3
pandas==2.3.3
pycountry==24.6.1
pydantic==1.10.12
python-dotenv==1.0.1
pytz==2024.1
regex==2026.1.15
rich==13.3.5
tinydb==4.8.0
tinydb-serialization==2.1.1
#tkinter==8.6.12
typer==0.21.1
typing_extensions==4.9.0
